Split consumers into base and demagog ones
