+++ apps/tests/test_consumer.py
@@ -4,7 +4,7 @@ import responses
 from django.test import TestCase
 from parameterized import parameterized
 
-from apps.publisher.consumers import JSONConsumer
+from ..consumers import JSONConsumer
 
 TEST_URL = 'http://i-test-you-all.org'
 OTHER_URL = 'http://i-dont-test-anything.org'
